Refactor components into folder
* app folder should be nearly empty.
* Avoid usage of relative imports
