+++ src/app/records/page.tsx
@@ -1,8 +1,8 @@
 'use client';
-import { PatientRecordTable } from '@/app/records/PatientRecordTable';
 import { LoadingPage } from '@/components/LoadingPage';
 import { LoadingUI } from '@/components/LoadingUI';
 import { PatientSearchInput } from '@/components/PatientSearchbar';
+import { PatientRecordTable } from '@/components/records/PatientRecordTable';
 import { useLoadingState } from '@/hooks/useLoadingState';
 import { Patient } from '@/types/Patient';
 import { Suspense, useState } from 'react';
