+++ src/app/records/patient-record/page.tsx
@@ -1,12 +1,12 @@
 'use server';
+import { EditPatient } from '@/components/records/patient/EditPatient';
 import { PatientInfoDetailSection } from '@/components/records/patient/PatientInfoDetailSection';
 import { PatientInfoHeaderSection } from '@/components/records/patient/PatientInfoHeaderSection';
-import { PrescriptionConsultCol } from '@/components/records/VitalPresConsultCol';
+import { PrescriptionConsultCol } from '@/components/records/PrescriptionConsultCol';
+import { ViewVital } from '@/components/records/vital/ViewVital';
 import { getConsultByVisitId } from '@/data/consult/getConsult';
 import { getPatientById } from '@/data/patient/getPatient';
 import { getVitalByVisit } from '@/data/vital/getVital';
-import { EditPatient } from './EditPatient';
-import { ViewVital } from './ViewVital';
 
 export default async function PatientRecordPage({
   searchParams,
