+++ src/components/Button.tsx
@@ -1,6 +1,6 @@
 'use client';
+import { LoadingUI } from '@/components/LoadingUI';
 import { ReactNode, useEffect, useState } from 'react';
-import { LoadingUI } from './LoadingUI';
 
 export function Button({
   text,
