+++ src/components/pharmacy/stock/AddMedicationModal.tsx
