+++ src/components/records/patient/PatientInfo.tsx
@@ -1,7 +1,7 @@
+import { PatientInfoActionBar } from '@/components/records/patient/PatientInfoActionBar';
+import { PatientInfoDetailSection } from '@/components/records/patient/PatientInfoDetailSection';
+import { PatientInfoHeaderSection } from '@/components/records/patient/PatientInfoHeaderSection';
 import { Patient } from '@/types/Patient';
-import { PatientInfoActionBar } from './PatientInfoActionBar';
-import { PatientInfoDetailSection } from './PatientInfoDetailSection';
-import { PatientInfoHeaderSection } from './PatientInfoHeaderSection';
 
 export function PatientInfo({ patient }: { patient: Patient | null }) {
   if (patient == null) {
