+++ src/components/records/vital/ViewVital.tsx
