+++ src/components/referrals/PatientSearchbarByReferral.tsx
@@ -1,4 +1,5 @@
 'use client';
+import { VillageOptionDropdown } from '@/components/VillageOptionDropdown';
 import { VillageContext } from '@/context/VillageContext';
 import { WithLoadingType } from '@/hooks/useLoadingState';
 import { VillagePrefix } from '@/types/VillagePrefixEnum';
@@ -11,7 +12,6 @@ import {
   useState,
 } from 'react';
 import { useDebouncedCallback } from 'use-debounce';
-import { VillageOptionDropdown } from '../VillageOptionDropdown';
 
 import {
   getReferrals,
