+++ src/components/registration/NewPatientModal.tsx
