+++ src/components/registration/PatientScanForm.tsx
